+++ src/api/digitalAsset.ts
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import {
+  AnyNumber,
   GetCollectionDataResponse,
   GetCurrentTokenOwnershipResponse,
   GetOwnedTokensResponse,
@@ -11,7 +12,6 @@ import {
   PaginationArgs,
   TokenStandard,
 } from "../types";
-import { AptosConfig } from "./aptosConfig";
 import { Account, AccountAddressInput } from "../core";
 import { InputGenerateTransactionOptions, SingleSignerTransaction } from "../transactions/types";
 import {
@@ -25,16 +25,15 @@ import {
   getTokenData,
   mintTokenTransaction,
 } from "../internal/digitalAsset";
+import { ProcessorType } from "../utils/const";
+import { AptosConfig } from "./aptosConfig";
+import { waitForIndexerOnVersion } from "./utils";
 
 /**
  * A class to query all `DigitalAsset` related queries on Aptos.
  */
 export class DigitalAsset {
-  readonly config: AptosConfig;
-
-  constructor(config: AptosConfig) {
-    this.config = config;
-  }
+  constructor(readonly config: AptosConfig) {}
 
   /**
    * Creates a new collection within the specified account
@@ -81,16 +80,23 @@ export class DigitalAsset {
    *
    * @param args.creatorAddress the address of the collection's creator
    * @param args.collectionName the name of the collection
+   * @param args.minimumLedgerVersion Optional ledger version to sync up to, before querying
    * @param args.options.tokenStandard the token standard to query
    * @returns GetCollectionDataResponse response type
    */
   async getCollectionData(args: {
     creatorAddress: AccountAddressInput;
     collectionName: string;
+    minimumLedgerVersion?: AnyNumber;
     options?: {
       tokenStandard?: TokenStandard;
     };
   }): Promise<GetCollectionDataResponse> {
+    await waitForIndexerOnVersion({
+      config: this.config,
+      minimumLedgerVersion: args.minimumLedgerVersion,
+      processorTypes: getTokenProcessorTypes(args.options?.tokenStandard),
+    });
     return getCollectionData({ aptosConfig: this.config, ...args });
   }
 
@@ -102,16 +108,23 @@ export class DigitalAsset {
    *
    * @param args.creatorAddress the address of the collection's creator
    * @param args.collectionName the name of the collection
+   * @param args.minimumLedgerVersion Optional ledger version to sync up to, before querying
    * @param args.options.tokenStandard the token standard to query
    * @returns the collection id
    */
   async getCollectionId(args: {
     creatorAddress: AccountAddressInput;
     collectionName: string;
+    minimumLedgerVersion?: AnyNumber;
     options?: {
       tokenStandard?: TokenStandard;
     };
   }): Promise<string> {
+    await waitForIndexerOnVersion({
+      config: this.config,
+      minimumLedgerVersion: args.minimumLedgerVersion,
+      processorTypes: getTokenProcessorTypes(args.options?.tokenStandard),
+    });
     return getCollectionId({ aptosConfig: this.config, ...args });
   }
 
@@ -141,9 +154,19 @@ export class DigitalAsset {
    * Gets token data given the address of a token.
    *
    * @param args.tokenAddress The address of the token
+   * @param args.minimumLedgerVersion Optional ledger version to sync up to, before querying
    * @returns GetTokenDataResponse containing relevant data to the token.
    */
-  async getTokenData(args: { tokenAddress: AccountAddressInput }): Promise<GetTokenDataResponse> {
+  async getTokenData(args: {
+    tokenAddress: AccountAddressInput;
+    minimumLedgerVersion?: AnyNumber;
+  }): Promise<GetTokenDataResponse> {
+    await waitForIndexerOnVersion({
+      config: this.config,
+      minimumLedgerVersion: args.minimumLedgerVersion,
+      // TODO(greg): Should take in a consistent input for token queries
+      processorTypes: getTokenProcessorTypes(undefined),
+    });
     return getTokenData({ aptosConfig: this.config, ...args });
   }
 
@@ -151,11 +174,19 @@ export class DigitalAsset {
    * Gets token ownership data given the address of a token.
    *
    * @param args.tokenAddress The address of the token
+   * @param args.minimumLedgerVersion Optional ledger version to sync up to, before querying
    * @returns GetCurrentTokenOwnershipResponse containing relevant ownership data of the token.
    */
   async getCurrentTokenOwnership(args: {
     tokenAddress: AccountAddressInput;
+    minimumLedgerVersion?: AnyNumber;
   }): Promise<GetCurrentTokenOwnershipResponse> {
+    await waitForIndexerOnVersion({
+      config: this.config,
+      minimumLedgerVersion: args.minimumLedgerVersion,
+      // TODO(greg): Should take in a consistent input for token queries
+      processorTypes: getTokenProcessorTypes(undefined),
+    });
     return getCurrentTokenOwnership({ aptosConfig: this.config, ...args });
   }
 
@@ -163,15 +194,23 @@ export class DigitalAsset {
    * Gets the tokens that the given address owns.
    *
    * @param args.ownerAddress The address of the owner
+   * @param args.minimumLedgerVersion Optional ledger version to sync up to, before querying
    * @returns GetOwnedTokensResponse containing ownership data of the tokens belonging to the ownerAddresss.
    */
   async getOwnedTokens(args: {
     ownerAddress: AccountAddressInput;
+    minimumLedgerVersion?: AnyNumber;
     options?: {
       pagination?: PaginationArgs;
       orderBy?: OrderBy<GetOwnedTokensResponse[0]>;
     };
   }): Promise<GetOwnedTokensResponse> {
+    await waitForIndexerOnVersion({
+      config: this.config,
+      minimumLedgerVersion: args.minimumLedgerVersion,
+      // TODO(greg): Should take in a consistent input for token queries
+      processorTypes: getTokenProcessorTypes(undefined),
+    });
     return getOwnedTokens({ aptosConfig: this.config, ...args });
   }
 
@@ -179,15 +218,35 @@ export class DigitalAsset {
    * Gets the activity data given the address of a token.
    *
    * @param args.tokenAddress The address of the token
+   * @param args.minimumLedgerVersion Optional ledger version to sync up to, before querying
    * @returns GetTokenActivityResponse containing relevant activity data to the token.
    */
   async getTokenActivity(args: {
     tokenAddress: AccountAddressInput;
+    minimumLedgerVersion?: AnyNumber;
     options?: {
       pagination?: PaginationArgs;
       orderBy?: OrderBy<GetTokenActivityResponse[0]>;
     };
   }): Promise<GetTokenActivityResponse> {
+    await waitForIndexerOnVersion({
+      config: this.config,
+      minimumLedgerVersion: args.minimumLedgerVersion,
+      // TODO(greg): Should take in a consistent input for token queries
+      processorTypes: getTokenProcessorTypes(undefined),
+    });
     return getTokenActivity({ aptosConfig: this.config, ...args });
+  }
+}
+
+function getTokenProcessorTypes(tokenStandard?: TokenStandard) {
+  switch (tokenStandard) {
+    case "v1":
+      return [ProcessorType.TOKEN_PROCESSOR];
+    case "v2":
+      return [ProcessorType.TOKEN_V2_PROCESSOR];
+    default:
+      // If it's something we don't recognize, or undefined, just do both
+      return [ProcessorType.TOKEN_PROCESSOR, ProcessorType.TOKEN_V2_PROCESSOR];
   }
 }