+++ src/api/event.ts
@@ -1,34 +1,39 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-import { AptosConfig } from "./aptosConfig";
 import { getAccountEventsByCreationNumber, getAccountEventsByEventType, getEvents } from "../internal/event";
 import { AnyNumber, GetEventsResponse, MoveStructId, OrderBy, PaginationArgs } from "../types";
 import { EventsBoolExp } from "../types/generated/types";
 import { AccountAddressInput } from "../core";
+import { ProcessorType } from "../utils/const";
+import { AptosConfig } from "./aptosConfig";
+import { waitForIndexerOnVersion } from "./utils";
 
 /**
  * A class to query all `Event` Aptos related queries
  */
 export class Event {
-  readonly config: AptosConfig;
-
-  constructor(config: AptosConfig) {
-    this.config = config;
-  }
+  constructor(readonly config: AptosConfig) {}
 
   /**
    * Get events by creation number and an account address
    *
    * @param args.accountAddress - The account address
    * @param args.creationNumber - The event creation number
+   * @param args.minimumLedgerVersion Optional ledger version to sync up to, before querying
    *
    * @returns Promise<GetEventsResponse>
    */
   async getAccountEventsByCreationNumber(args: {
     accountAddress: AccountAddressInput;
     creationNumber: AnyNumber;
+    minimumLedgerVersion?: AnyNumber;
   }): Promise<GetEventsResponse> {
+    await waitForIndexerOnVersion({
+      config: this.config,
+      minimumLedgerVersion: args.minimumLedgerVersion,
+      processorTypes: [ProcessorType.EVENTS_PROCESSOR],
+    });
     return getAccountEventsByCreationNumber({ aptosConfig: this.config, ...args });
   }
 
@@ -37,17 +42,24 @@ export class Event {
    *
    * @param args.accountAddress - The account address
    * @param args.eventType - The event type
+   * @param args.minimumLedgerVersion Optional ledger version to sync up to, before querying
    *
    * @returns Promise<GetEventsResponse>
    */
   async getAccountEventsByEventType(args: {
     accountAddress: AccountAddressInput;
     eventType: MoveStructId;
+    minimumLedgerVersion?: AnyNumber;
     options?: {
       pagination?: PaginationArgs;
       orderBy?: OrderBy<GetEventsResponse[0]>;
     };
   }): Promise<GetEventsResponse> {
+    await waitForIndexerOnVersion({
+      config: this.config,
+      minimumLedgerVersion: args.minimumLedgerVersion,
+      processorTypes: [ProcessorType.EVENTS_PROCESSOR],
+    });
     return getAccountEventsByEventType({ aptosConfig: this.config, ...args });
   }
 
@@ -55,7 +67,7 @@ export class Event {
    * Get all events
    *
    * An optional `where` can be passed in to filter out the response.
-   *
+   *@param args.minimumLedgerVersion Optional ledger version to sync up to, before querying
    * @example
    * ```
    * { where:
@@ -68,12 +80,18 @@ export class Event {
    * @returns GetEventsQuery response type
    */
   async getEvents(args?: {
+    minimumLedgerVersion?: AnyNumber;
     options?: {
       where?: EventsBoolExp;
       pagination?: PaginationArgs;
       orderBy?: OrderBy<GetEventsResponse[0]>;
     };
   }): Promise<GetEventsResponse> {
+    await waitForIndexerOnVersion({
+      config: this.config,
+      minimumLedgerVersion: args?.minimumLedgerVersion,
+      processorTypes: [ProcessorType.EVENTS_PROCESSOR],
+    });
     return getEvents({ aptosConfig: this.config, ...args });
   }
 }