+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 # Unreleased
 
 - Respect `API_KEY` option in `clientConfig` when making indexer and/or fullnode queries
+- [`Added`] Added `waitForIndexer` function to wait for indexer to sync up with full node. All indexer query functions now accepts a new optional param `minimumLedgerVersion` to wait for indexer to sync up with the target processor.
 
 Breaking:
 - Changes ANS date usage to consistently use epoch timestamps represented in milliseconds.